Переносит код из main.js в соответствующие модули. Подключает к main.js модули для дальнейшей работы
