+++ js/browse-photos.js
@@ -1,0 +1,8 @@
+import { checkStringLength } from './util.js';
+
+function checkBrowsePhoto () {
+  const check = checkStringLength('check', 5);
+  return check;
+}
+
+export {checkBrowsePhoto};
