+++ js/effects-photo.js
@@ -1,0 +1,8 @@
+import { checkStringLength } from './util.js';
+
+function checkEffectsPhoto () {
+  const check = checkStringLength('check', 5);
+  return check;
+}
+
+export {checkEffectsPhoto};
