+++ js/representation-filter.js
@@ -1,0 +1,8 @@
+import { checkStringLength } from './util.js';
+
+function checkRepresentationFilter () {
+  const check = checkStringLength('check', 5);
+  return check;
+}
+
+export {checkRepresentationFilter};
