+++ js/scale.js
@@ -1,0 +1,8 @@
+import { checkStringLength } from './util.js';
+
+function checkScale () {
+  const check = checkStringLength('check', 5);
+  return check;
+}
+
+export {checkScale};
